+++ src/main/java/oop/project/deserializers/AuthDeserializer.java
@@ -1,0 +1,7 @@
+package oop.project.deserializers;
+
+public class AuthDeserializer {
+    public int status_code;
+    public String status_message;
+    public String success;
+}
