+++ src/main/java/oop/project/deserializers/CreditsDeserializer.java
@@ -1,0 +1,9 @@
+package oop.project.deserializers;
+
+import java.util.List;
+
+public class CreditsDeserializer {
+    public List<FilmDeserializer> cast;
+    public List<FilmDeserializer> crew;
+    public int id;
+}
