+++ src/main/java/oop/project/deserializers/FilmDeserializer.java
@@ -1,0 +1,34 @@
+package oop.project.deserializers;
+
+import java.util.List;
+
+public class FilmDeserializer {
+    public boolean adult;
+    public String backdrop_path;
+    public int budget;
+    public List<Integer> genre_ids;
+    public String homepage;
+    public int id;
+    public String imdb_id;
+    public String media_type;
+    public String[] origin_country;
+    public String original_language;
+    public String original_title;
+    public String overview;
+    public double popularity;
+    public String poster_path;
+    public String release_date;
+    public String revenue;
+    public String runtime;
+    public List<LanguageDeserializer> spoken_languages;
+    public String status;
+    public String tagline;
+    public String title;
+    public boolean video;
+    public double vote_average;
+    public double vote_count;
+//    public boolean belongs_to_collection;
+//    public List<Integer> genres;
+//    public List<FilmDeserializer> production_companies;
+//    public List<FilmDeserializer> production_countries;
+}
