+++ src/main/java/oop/project/deserializers/GenreDeserializer.java
@@ -1,0 +1,6 @@
+package oop.project.deserializers;
+
+public class GenreDeserializer {
+    public int id;
+    public String name;
+}
