+++ src/main/java/oop/project/deserializers/LanguageDeserializer.java
@@ -1,0 +1,7 @@
+package oop.project.deserializers;
+
+public class LanguageDeserializer {
+    public String english_name;
+    public String iso_639_1;
+    public String name;
+}
