+++ src/main/java/oop/project/shared/CommandWaiter.java
@@ -1,0 +1,13 @@
+package oop.project.shared;
+
+public enum CommandWaiter {
+    NONE,
+    GENRE,
+    YEAR,
+    MOVIESEARCH,
+    ACTORSEARCH,
+    SIMILAR,
+    RECOMMENDED,
+    FINDBYID,
+    SETAGE;
+}
