+++ src/main/java/oop/project/validators/IdValidator.java
@@ -1,0 +1,19 @@
+package oop.project.validators;
+
+public class IdValidator implements Validator<String> {
+    @Override
+    public boolean isValid(String input) {
+        try {
+            Integer.parseInt(input);
+            return true;
+        } catch (NumberFormatException e) {
+            return false;
+        }
+    }
+
+    @Override
+    public String getErrorMessage() {
+        return "Пожалуйста, введите корректный ID!";
+    }
+}
+
