+++ src/main/java/oop/project/validators/Validator.java
@@ -1,0 +1,7 @@
+package oop.project.validators;
+
+public interface Validator<T> {
+    boolean isValid(T input);
+    String getErrorMessage();
+}
+
