--- src/main/java/org/oopproject/deserializers/AuthDeserializer.java
@@ -1,7 +1,0 @@
-package org.oopproject.deserializers;
-
-public class AuthDeserializer {
-    public int status_code;
-    public String status_message;
-    public String success;
-}
