--- src/main/java/org/oopproject/deserializers/CreditsDeserializer.java
@@ -1,9 +1,0 @@
-package org.oopproject.deserializers;
-
-import java.util.List;
-
-public class CreditsDeserializer {
-    public List<FilmDeserializer> cast;
-    public List<FilmDeserializer> crew;
-    public int id;
-}
