--- src/main/java/org/oopproject/deserializers/DateDeserializer.java
@@ -1,22 +1,0 @@
-package org.oopproject.deserializers;
-
-public class DateDeserializer {
-    private String maximum;
-    private String minimum;
-
-    public String getMaximum() {
-        return maximum;
-    }
-
-    public void setMaximum(String maximum) {
-        this.maximum = maximum;
-    }
-
-    public String getMinimum() {
-        return minimum;
-    }
-
-    public void setMinimum(String minimum) {
-        this.minimum = minimum;
-    }
-}
