--- src/main/java/org/oopproject/deserializers/GenreDeserializer.java
@@ -1,6 +1,0 @@
-package org.oopproject.deserializers;
-
-public class GenreDeserializer {
-    public int id;
-    public String name;
-}
