--- src/main/java/org/oopproject/deserializers/LanguageDeserializer.java
@@ -1,7 +1,0 @@
-package org.oopproject.deserializers;
-
-public class LanguageDeserializer {
-    public String english_name;
-    public String iso_639_1;
-    public String name;
-}
