--- src/main/java/org/oopproject/utils/CommandWaiter.java
@@ -1,13 +1,0 @@
-package org.oopproject.utils;
-
-public enum CommandWaiter {
-    NONE,
-    GENRE,
-    YEAR,
-    MOVIESEARCH,
-    ACTORSEARCH,
-    SIMILAR,
-    RECOMMENDED,
-    FINDBYID,
-    SETAGE;
-}
