--- src/main/java/org/oopproject/validators/IdValidator.java
@@ -1,19 +1,0 @@
-package org.oopproject.validators;
-
-public class IdValidator implements Validator<String> {
-    @Override
-    public boolean isValid(String input) {
-        try {
-            Integer.parseInt(input);
-            return true;
-        } catch (NumberFormatException e) {
-            return false;
-        }
-    }
-
-    @Override
-    public String getErrorMessage() {
-        return "Пожалуйста, введите корректный ID!";
-    }
-}
-
