--- src/main/java/org/oopproject/validators/Validator.java
@@ -1,7 +1,0 @@
-package org.oopproject.validators;
-
-public interface Validator<T> {
-    boolean isValid(T input);
-    String getErrorMessage();
-}
-
