--- src/main/java/org/oopproject/validators/YearValidator.java
@@ -1,22 +1,0 @@
-package org.oopproject.validators;
-
-public class YearValidator implements Validator<String> {
-    private final int minYear = 1895;
-    private final int maxYear = java.time.Year.now().getValue();
-
-    @Override
-    public boolean isValid(String input) {
-        try {
-            int year = Integer.parseInt(input);
-            return year >= minYear && year <= maxYear;
-        } catch (NumberFormatException e) {
-            return false;
-        }
-    }
-
-    @Override
-    public String getErrorMessage() {
-        return "Пожалуйста, введите год в диапазоне от " + minYear + " до " + maxYear;
-    }
-}
-
